refactor(App): removed everything related to st-cuts
