--- src/lib/helpers/addSTCuts/index.ts
@@ -1,1 +1,0 @@
-export { addSTCuts } from "./addSTCuts";
